extract trackmeta
